pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "explot"
version = "0.1.0"
description = "Universal data analysis pipeline for tabular datasets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "jinja2>=3.1",
  "numpy>=1.26",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "scikit-learn>=1.4",
  "scipy>=1.12",
]

[project.optional-dependencies]
autoencoder = [
  "torch>=2.2",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
  "torch>=2.2",
]
report = [
  "plotly>=5.20",
]
excel = [
  "openpyxl>=3.1",
]
parquet = [
  "pyarrow>=15.0",
]

[project.scripts]
explot = "explot.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["explot*", "simulator*"]

[tool.setuptools.package-data]
explot = [
  "stages/manifest.yaml",
  "stages/*/stage.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
norecursedirs = ["tests/_tmp", "_tmp_test_cases"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
